Implement basic BLE GATT server connection logic.
The general logic has been implemented to get a connection to a BLE GATT server. The UI still needs to be frozen while attempting to connect and there needs to be some way to make sure that the connection is completely disconnected. Otherwise, the GATT server implementation needs figured out.